+++ src/server/core_server.js
@@ -5,6 +5,7 @@ var opts = require('commander');
 var winston = require('winston');
 var configurator = require('../../config');
 var file = require('./api/v3/file.js');
+var fs = require('fs');
 
 /******************************************************************************/
 
@@ -42,18 +43,41 @@ function CoreServer() {
   this.port = opts.port || this.config.port || 8080;
 
   // set up machine tool option
-  if (opts.tool) {
-    this.machinetool = opts.tool;
-  } else {
-    this.machinetool = this.config.machine.dir;
+  if(opts.tool){
+    try{
+      fs.accessSync(opts.tool, fs.R_OK);
+      this.machinetool = opts.tool;
+    }
+    catch(err){
+      console.log('Couldn\'t Access Path Provided For Machine, Ending...');
+      this.machinetool = null;
+      process.exit();
+    }
   }
-
+  else if(this.config.machine !== undefined && this.config.machine.dir !== undefined){
+    if(this.config.machine.dir.trim() !== ""){
+      try{
+        fs.accessSync(this.config.machine.dir, fs.R_OK);
+        this.machinetool = this.config.machine.dir;
+      }
+      catch(err){
+        console.log('Couldn\'t Access Path Provided For Machine, Ending...');
+        this.machinetool = null;
+        process.exit();
+      }
+    }
+    else{        
+      this.machinetool = null;
+    }
+  }
+  else{
+    this.machinetool = null;
+  }
   // set up filepath option
-  if (opts.file) {
-    this.project = opts.file;
-  } else {
-    this.project = this.config.file.dir;
-  }
+  if(opts.file)
+      this.project = opts.file;
+  else
+      this.project = this.config.file.dir;
   file.init(this.project, this.machinetool);
 
   // Establish core